created separate files for weeklyschedule class and calculateschedule function
